// tools/config.d.ts
export declare const config: {
    ownerPhoneNumber: string;
    calendarId: string;
    serviceAccountPath: string;
    ollamaBaseUrl: string;
    ollamaModel: string;
    timezone: string;
    whitelistJids: string[];
    dbPath: string;
    imapAccounts: Record<string, {
        host: string;
        port: number;
        user: string;
        password: string;
    }>;
    piperBinaryPath: string;
    piperVoiceModel: string;
    ttsOutputDir: string;
    dashboardPort: number;
    dashboardToken: string;
    immichBaseUrl: string;
    immichApiKey: string;
};
//# sourceMappingURL=config.d.ts.map